Fixed a wrong type declaration which I just happened to notice in the #else section of an #if NhlNeedProto directive. This code is probably never going to be used, but just in case...
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10115 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/SrcTree.c
@@ -1,6 +1,6 @@
 
 /*
- *      $Id: SrcTree.c,v 1.40 2003-06-10 00:20:22 dbrown Exp $
+ *      $Id: SrcTree.c,v 1.41 2008-04-04 23:26:28 dbrown Exp $
  */
 /************************************************************************
 *									*
@@ -1719,7 +1719,7 @@ void * _NclMakeRealExpr
 (double real,char *string_rep)
 #else
 (real,string_rep)
-float real;
+double real;
 char *string_rep;
 #endif
 {
